requirements.txt
python-dotenv
beautifulsoup4
pytest
mdformat
langchain-core
langchain-text-splitters
langchain-community
tenacity
qdrant-client
httpx
fastembed
fastmcp
faker
sentence-transformers
mistralai
google-generativeai